The PieChart visualizations data are now being dynamically built
Before this the community names were all hardcoded, hardcoded seven
communities, the ordering of the query was always assumed to be the same,
and the ordering of the first query was assumed to be the same as the
second query.

Now generating the data dynamically given the number of communities
that actually exist. And use the popup counts as key/value pairs.